# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Copilot AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CarbonEstimateInput {
  /** The user action to estimate (e.g. "fly from London to Paris") */
  query: string;
  /**
     * Optional category hint
     * @nullable
     */
  category?: string | null;
}

export interface Alternative {
  label: string;
  co2Kg: number;
  reductionPercent: number;
  /** @nullable */
  extraTimeMinutes?: number | null;
  /** @nullable */
  moneySavedUsd?: number | null;
  /** @nullable */
  description?: string | null;
}

export interface CarbonEstimateResult {
  action: string;
  category: string;
  co2Kg: number;
  explanation: string;
  alternatives: Alternative[];
  /** 0-1 confidence score */
  confidenceScore: number;
  /** Key factors contributing to emissions */
  factors?: string[];
}

export interface CarbonEntryInput {
  action: string;
  category: string;
  co2Kg: number;
  /**
     * JSON-encoded alternatives
     * @nullable
     */
  alternatives?: string | null;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  chosenAlternative?: string | null;
  /** @nullable */
  savedCo2Kg?: number | null;
}

export interface CarbonEntry {
  id: number;
  action: string;
  category: string;
  co2Kg: number;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  alternatives?: string | null;
  /** @nullable */
  chosenAlternative?: string | null;
  /** @nullable */
  savedCo2Kg?: number | null;
  createdAt: string;
}

export interface CarbonStats {
  totalCo2Kg: number;
  savedCo2Kg: number;
  totalEntries: number;
  treesEquivalent: number;
  phonesCharged: number;
  homesEquivalent?: number;
  streakDays: number;
  moneySavedUsd?: number;
}

export interface CategoryBreakdown {
  category: string;
  co2Kg: number;
  count: number;
  percentage: number;
}

export type UserProfileDietPreference = typeof UserProfileDietPreference[keyof typeof UserProfileDietPreference];


export const UserProfileDietPreference = {
  vegan: 'vegan',
  vegetarian: 'vegetarian',
  flexitarian: 'flexitarian',
  omnivore: 'omnivore',
} as const;

export type UserProfileBudgetSensitivity = typeof UserProfileBudgetSensitivity[keyof typeof UserProfileBudgetSensitivity];


export const UserProfileBudgetSensitivity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type UserProfileTimeSensitivity = typeof UserProfileTimeSensitivity[keyof typeof UserProfileTimeSensitivity];


export const UserProfileTimeSensitivity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface UserProfile {
  id: number;
  homeCity: string;
  dietPreference: UserProfileDietPreference;
  hasVehicle?: boolean;
  budgetSensitivity: UserProfileBudgetSensitivity;
  timeSensitivity: UserProfileTimeSensitivity;
  /** @nullable */
  typicalCommute?: string | null;
  /** @nullable */
  monthlyBudget?: number | null;
  createdAt?: string;
}

export type UserProfileUpdateDietPreference = typeof UserProfileUpdateDietPreference[keyof typeof UserProfileUpdateDietPreference];


export const UserProfileUpdateDietPreference = {
  vegan: 'vegan',
  vegetarian: 'vegetarian',
  flexitarian: 'flexitarian',
  omnivore: 'omnivore',
} as const;

export type UserProfileUpdateBudgetSensitivity = typeof UserProfileUpdateBudgetSensitivity[keyof typeof UserProfileUpdateBudgetSensitivity];


export const UserProfileUpdateBudgetSensitivity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type UserProfileUpdateTimeSensitivity = typeof UserProfileUpdateTimeSensitivity[keyof typeof UserProfileUpdateTimeSensitivity];


export const UserProfileUpdateTimeSensitivity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface UserProfileUpdate {
  homeCity?: string;
  dietPreference?: UserProfileUpdateDietPreference;
  hasVehicle?: boolean;
  budgetSensitivity?: UserProfileUpdateBudgetSensitivity;
  timeSensitivity?: UserProfileUpdateTimeSensitivity;
  typicalCommute?: string;
  monthlyBudget?: number;
}

export interface CarbonScores {
  daily: number;
  weekly: number;
  monthly: number;
  yearly: number;
  dailyBudget: number;
  weeklyChange?: number;
  monthlyChange?: number;
}

export interface ScoreHistoryPoint {
  date: string;
  co2Kg: number;
  savedCo2Kg?: number;
}

export interface Achievement {
  id: number;
  title: string;
  description: string;
  icon: string;
  isUnlocked: boolean;
  /** @nullable */
  unlockedAt: string | null;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  target?: number | null;
}

export type ListCarbonEntriesParams = {
limit?: number;
offset?: number;
category?: string;
};

export type GetRecentEntriesParams = {
limit?: number;
};

export type GetScoreHistoryParams = {
period?: GetScoreHistoryPeriod;
};

export type GetScoreHistoryPeriod = typeof GetScoreHistoryPeriod[keyof typeof GetScoreHistoryPeriod];


export const GetScoreHistoryPeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;