# lib/api-zod/src/generated/types/achievement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Copilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Achievement {
  id: number;
  title: string;
  description: string;
  icon: string;
  isUnlocked: boolean;
  /** @nullable */
  unlockedAt: Date | null;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  target?: number | null;
}