# lib/api-zod/src/generated/types/alternative.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Copilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Alternative {
  label: string;
  co2Kg: number;
  reductionPercent: number;
  /** @nullable */
  extraTimeMinutes?: number | null;
  /** @nullable */
  moneySavedUsd?: number | null;
  /** @nullable */
  description?: string | null;
}