# lib/api-zod/src/generated/types/carbonEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Copilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CarbonEntry {
  id: number;
  action: string;
  category: string;
  co2Kg: number;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  alternatives?: string | null;
  /** @nullable */
  chosenAlternative?: string | null;
  /** @nullable */
  savedCo2Kg?: number | null;
  createdAt: Date;
}