# lib/api-zod/src/generated/types/carbonEstimateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Copilot AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Alternative } from './alternative';

export interface CarbonEstimateResult {
  action: string;
  category: string;
  co2Kg: number;
  explanation: string;
  alternatives: Alternative[];
  /** 0-1 confidence score */
  confidenceScore: number;
  /** Key factors contributing to emissions */
  factors?: string[];
}