# lib/api-zod/src/generated/types/carbonScores.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Copilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CarbonScores {
  daily: number;
  weekly: number;
  monthly: number;
  yearly: number;
  dailyBudget: number;
  weeklyChange?: number;
  monthlyChange?: number;
}