# lib/api-zod/src/generated/types/scoreHistoryPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Copilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreHistoryPoint {
  date: string;
  co2Kg: number;
  savedCo2Kg?: number;
}