# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Copilot AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileBudgetSensitivity } from './userProfileBudgetSensitivity';
import type { UserProfileDietPreference } from './userProfileDietPreference';
import type { UserProfileTimeSensitivity } from './userProfileTimeSensitivity';

export interface UserProfile {
  id: number;
  homeCity: string;
  dietPreference: UserProfileDietPreference;
  hasVehicle?: boolean;
  budgetSensitivity: UserProfileBudgetSensitivity;
  timeSensitivity: UserProfileTimeSensitivity;
  /** @nullable */
  typicalCommute?: string | null;
  /** @nullable */
  monthlyBudget?: number | null;
  createdAt?: Date;
}