# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Copilot AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileUpdateBudgetSensitivity } from './userProfileUpdateBudgetSensitivity';
import type { UserProfileUpdateDietPreference } from './userProfileUpdateDietPreference';
import type { UserProfileUpdateTimeSensitivity } from './userProfileUpdateTimeSensitivity';

export interface UserProfileUpdate {
  homeCity?: string;
  dietPreference?: UserProfileUpdateDietPreference;
  hasVehicle?: boolean;
  budgetSensitivity?: UserProfileUpdateBudgetSensitivity;
  timeSensitivity?: UserProfileUpdateTimeSensitivity;
  typicalCommute?: string;
  monthlyBudget?: number;
}